# src/app/api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Mission Control API",
    "description": "AI Agent Orchestration Dashboard — REST API for managing agents, tasks, sessions, tokens, webhooks, and more.",
    "version": "1.2.0",
    "license": { "name": "MIT" },
    "contact": { "name": "Mission Control", "url": "https://github.com/builderz-labs/mission-control" }
  },
  "servers": [
    { "url": "/", "description": "Current server" }
  ],
  "security": [
    { "SessionCookie": [] },
    { "ApiKey": [] }
  ],
  "components": {
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "mc-session",
        "description": "Session token obtained via /api/auth/login"
      },
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key (env var or DB-backed managed token)"
      }
    },
...
  },
  "paths": {
...
    "/api/status": {
      "get": {
        "tags": ["System"],
        "summary": "Get system and application status",
        "description": "Returns system health (uptime, memory, disk), dashboard stats, gateway status, and available models.",
        "parameters": [{ "name": "section", "in": "query", "schema": { "type": "string", "enum": ["system", "dashboard", "gateway", "models", "health"] } }],
        "responses": { "200": { "description": "Status data" } }
      }
    },
...
  },
...
}